refactor(iwork): move the Pages backend alongside the other backends
Addresses review feedback on #4062.

`docling/backend` holds backend modules, so the backend itself belongs
there rather than inside a supporting package. `IWorkPagesDocumentBackend`
moves back to `docling/backend/iwork_backend.py`, which is also where the
Numbers and Keynote backends will go when they follow, matching how
`opendocument_backend.py` holds all three OpenDocument backends.

`docling/backend/iwork` keeps only what those backends are built from: the
IWA container reader, the content model, and the two readers that fill it.
The package no longer re-exports the backend, which would now be a cycle.
