+++ docling/backend/iwork/__init__.py
@@ -1,14 +1,12 @@
 # SPDX-FileCopyrightText: The Docling Contributors
 # SPDX-License-Identifier: MIT
 
-"""Support for Apple iWork documents.
+"""Supporting modules for the Apple iWork backends.
+
+The backends themselves live in :mod:`docling.backend.iwork_backend`, alongside
+the other document backends; this package holds what they are built from.
 
 ``iwa`` reads the IWA container that Pages, Numbers and Keynote have used since
-2013. ``content`` models what a Pages document holds, ``pages_iwa`` and
-``pages_xml`` read the two container generations into that model, and
-``pages_backend`` turns the result into a ``DoclingDocument``.
+2013. ``content`` models what a Pages document holds, and ``pages_iwa`` and
+``pages_xml`` read the two container generations into that model.
 """
-
-from docling.backend.iwork.pages_backend import IWorkPagesDocumentBackend
-
-__all__ = ["IWorkPagesDocumentBackend"]
